Add files via upload
About Float and Clear
Float :- The float property is used for positioning and formatting content e.g. let an image float left to the text in a container.
Clear :- When we use the float property, and we want the next element below (not on right or left), we will have to use the clear property.